# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Crud API Documentation",
        "description": "Laravel CRUD API description",
        "contact": {
            "email": "mahbubhussaincse@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "ab015cc2055ac75e6d23c628675c26e7",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Per page count",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order By column name",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order ordering - asc or desc",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "Create product",
                "operationId": "7ff81b1681ff3180e5b73a2e15f2c67b",
                "requestBody": {
                    "description": "Product objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string",
                                        "example": "Product title"
                                    },
                                    "slug": {
                                        "description": "Product slug",
                                        "type": "string",
                                        "example": "product-title"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "image": {
                                        "description": "Product image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product detail",
                "description": "Get product detail",
                "operationId": "47affec446a4e21ba3cbf0b6b14abb44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product",
                "operationId": "914fbddeaed4fc2e831b9f5fb0b05b40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "description": "request method",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title",
                                    "price",
                                    "slug"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Product id",
                                        "type": "integer",
                                        "example": "Product id"
                                    },
                                    "title": {
                                        "description": "Product title",
                                        "type": "string",
                                        "example": "Product title"
                                    },
                                    "slug": {
                                        "description": "Product slug",
                                        "type": "string",
                                        "example": "product-title"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "image": {
                                        "description": "Product image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "description": "Delete product",
                "operationId": "33173a96c7ab5ed4ba9ceaffb4c6003d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}